+++ example/lib/main.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import 'screens/expandable_multiple_choice/expandable_multiple_choice_screen.dart';
+import 'screens/expandable_multiple_choice/expanded_screen.dart';
 import 'screens/expandable_single_choice/expandable_single_choice_screen.dart';
 import 'screens/stack/stack_screen.dart';
 import 'widgets/btn1.dart';
@@ -38,7 +38,7 @@ class MyApp extends StatelessWidget {
             ),
             divider,
             const Btn1(
-              screen: ExpandableMultipleChoiceScreen(),
+              screen: ExpandedScreen(),
               title: "Expanded variants",
             ),
             divider,