+++ example/lib/screens/expandable_multiple_choice/ex_expandable_tree_expanded_screen.dart
@@ -38,7 +38,7 @@ class _ExExpandableTreeExpandedScreenState
           children: [
             Expanded(
               flex: 4,
-              child: ExpandableTreeWidget(_tree),
+              child: ExpandedWidget(_tree),
             ),
             Expanded(
               flex: 1,