+++ example/lib/screens/expandable_single_choice/ex_tree_single_choice.dart
@@ -85,47 +85,40 @@ class _VTSNodeWidget extends StatefulWidget {
 }
 
 class _VTSNodeWidgetState<T extends AbsNodeType> extends State<_VTSNodeWidget>
-    with SingleTickerProviderStateMixin, ExpandableTreeMixin<EasyNodeType> {
-  final Tween<double> _turnsTween = Tween<double>(begin: -0.25, end: 0.0);
-
+    with SingleTickerProviderStateMixin, ExpandedWidgetMixin<EasyNodeType> {
   @override
-  initState() {
-    super.initState();
-    initTree();
-    initRotationController();
+  void onInit() {
+    tree = widget.tree;
+    rotationController = AnimationController(
+      duration: const Duration(milliseconds: 300),
+      vsync: this,
+    );
     if (tree.data.isExpanded) {
       rotationController.forward();
     }
   }
 
   @override
-  void initTree() {
-    tree = widget.tree;
-  }
-
-  @override
-  void initRotationController() {
-    rotationController = AnimationController(
-      duration: const Duration(milliseconds: 300),
-      vsync: this,
-    );
+  initState() {
+    super.initState();
+    onInit();
   }
 
   @override
   void dispose() {
-    super.disposeRotationController();
+    onDispose();
     super.dispose();
   }
 
   @override
-  Widget build(BuildContext context) => buildView();
+  Widget build(BuildContext context) => buildView(context);
 
   @override
-  Widget buildNode() {
+  Widget buildNode(BuildContext context) {
     if (!widget.tree.data.isShowedInSearching) return const SizedBox.shrink();
 
     return InkWell(
-      onTap: updateStateToggleExpansion,
+      onTap: updateToggleExpansion,
       child: Row(
         children: [
           buildRotationIcon(),
@@ -140,13 +133,13 @@ class _VTSNodeWidgetState<T extends AbsNodeType> extends State<_VTSNodeWidget>
 
   Widget buildRotationIcon() {
     return RotationTransition(
-      turns: _turnsTween.animate(rotationController),
+      turns: turnsTween.animate(rotationController),
       child: tree.isLeaf
           ? Container()
           : IconButton(
               iconSize: 16,
               icon: const Icon(Icons.expand_more, size: 16.0),
-              onPressed: updateStateToggleExpansion,
+              onPressed: updateToggleExpansion,
             ),
     );
   }
@@ -183,7 +176,7 @@ class _VTSNodeWidgetState<T extends AbsNodeType> extends State<_VTSNodeWidget>
   //* __________________________________________________________________________
 
   @override
-  List<Widget> generateChildrenNodesWidget(List<TreeType<EasyNodeType>> list) =>
+  List<Widget> genChildrenWidgets(List<TreeType<EasyNodeType>> list) =>
       List.generate(
         list.length,
         (int index) => _VTSNodeWidget(
@@ -193,5 +186,5 @@ class _VTSNodeWidgetState<T extends AbsNodeType> extends State<_VTSNodeWidget>
       );
 
   @override
-  void updateStateToggleExpansion() => setState(() => toggleExpansion());
+  void updateToggleExpansion() => setState(() => toggleExpansion());
 }